Remove useless class
