+++ modyn/models/resnet18/resnet18.py
@@ -1,7 +1,9 @@
 from typing import Any
 
-# torchvision resnet updated to add the embedding recorder.
-from modyn.models.resnet18.resnet_torchvision import ResNetModyn
+import torch
+from modyn.models.modyn_model import ModynModel
+from torch import Tensor, nn
+from torchvision.models.resnet import BasicBlock, ResNet
 
 
 class ResNet18:
@@ -9,3 +11,34 @@ class ResNet18:
     def __init__(self, model_configuration: dict[str, Any], device: str, amp: bool) -> None:
         self.model = ResNetModyn(model_configuration)
         self.model.to(device)
+
+
+# the following class is adapted from
+# torchvision https://github.com/pytorch/vision/blob/main/torchvision/models/resnet.py
+
+
+class ResNetModyn(ResNet, ModynModel):
+    def __init__(self, model_configuration: dict[str, Any]) -> None:
+        super().__init__(BasicBlock, [2, 2, 2, 2], **model_configuration)
+
+    def _forward_impl(self, x: Tensor) -> Tensor:
+        x = self.conv1(x)
+        x = self.bn1(x)
+        x = self.relu(x)
+        x = self.maxpool(x)
+
+        x = self.layer1(x)
+        x = self.layer2(x)
+        x = self.layer3(x)
+        x = self.layer4(x)
+
+        x = self.avgpool(x)
+        x = torch.flatten(x, 1)
+        # the following line is the only difference compared to the original implementation
+        x = self.embedding_recorder(x)
+        x = self.fc(x)
+
+        return x
+
+    def get_last_layer(self) -> nn.Module:
+        return self.fc
